requirements.txt
# ── Core data ────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0          # Parquet read/write

# ── Machine learning ─────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0             # Feature importance / explainability

# ── Configuration ────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── Visualisation ────────────────────────────────────────────────────────────
matplotlib>=3.7.0
Pillow>=10.0.0           # Dashboard image composition

# ── Model persistence ────────────────────────────────────────────────────────
joblib>=1.3.0

# ── Web app ──────────────────────────────────────────────────────────────────
streamlit>=1.28.0

# ── Testing & linting (dev) ──────────────────────────────────────────────────
pytest>=7.4.0
flake8>=6.1.0
